Corrected Logical errors in ResourceManager.cpp
* Removed comparison of -1 to unsigned int
* Corrected comparison of current_device to HID_INTERFACE_ANY
* Exposed downstream errors in Windows implementation of CM_MP750
controller corrected
* Added "Off" mode

+++ ResourceManager.h
@@ -42,8 +42,8 @@ typedef struct
     HIDDeviceDetectorFunction   function;
     unsigned int                address;
     int                         interface;
-    int                         usage_page;
-    int                         usage;
+    unsigned short              usage_page;
+    unsigned short              usage;
 } HIDDeviceDetectorBlock;
 
 typedef void (*DeviceListChangeCallback)(void *);
